TTSP.c ver. 0.3_g
Refactored all files to create a single small main in TTSP.c; encountered
a name conflict with the directory TSP, so renamed TSP.c to TTSP.c;
moved CitiesSmallSet into main, and used it to test the new refactored
code.  Refactored travellingSalesman.c to accept a struct City array.
Changed code so it did not try to report on a shortest distance if the
algorithm only displayed permutations; moved shortestRoute inside
doPermutations so it would not be a global variable.  Moved heapsAlg
from the minilab directory into the main directory.