nearestNeighbor.c ver 1.0_a Refactored nearestNeighbor
nearestNeighborSearch now checks for shortest distance by looking
at every city in first position and picks the path with the
shortest overall round-trip.
Added new procedure nn that checks one route with an initial city,
then returns the distance.
I discovered a bug in my original code; the exchange procedure was
not exchanging the city that produced the shortest route, but
rather was always exchanging the last city in the source list.
After this bug was fixed, nearestNeighbor now finds the shortest
distance in all cases that I can check with Heap's Algorithim
(4-12 cities).
Refactored travellingSalesman to not update indices.
Refactored ttsp.h to include new procedure nn, and use the
prototype for exchange.