kNN.c ver 0.1_d: Refactored to correctly recourse knn times
Was incorrectly recursing by decrementing knn; needed to recurse
knn times while leaving knn the same; needed to add copies at
each recurse, with copies of the size as well.  It seems as though
it is now recursing and producing identical nodes, but only for the
first few nodes.  Not sure what is happening.