TTSP.c ver 1.0_c Refactored kNN to recurse on first k of initial
city, not all cities, which was completely intractable.  Re-
factored TTSP.c to use the same code for NEAREST_NEIGHBOR and
KNN options as much as possible.  Changed reporting output a
bit.  No longer displays the routes if size is over 100.

kNN.c ver 0.1_e Refactored to recurse on first k neighbors of the
initial city in the list, not all of the cities as the list was
being built.  This produced a substantial improvement, and showed
that the k-NN algorithm successfully finds shorter routes up to
about k==3.  Refactored comments.  Made substantive change was at
lines 108 in the for loop, which limits the number of times the
routine recurses to k.  The loop was not looping properly because
knn was changing, so needed to add a local variable that would
stay the same.  After that fix, the algorithm showed that it
would find shorter routes with higher k, up to about k==3,at which
point it continued to find the same route distance.

nearestNeighbor.c ver 1.0_c Added a count variable for the number
of iterations performed.

ttsp.h ver 1.0_d Added count variable to struct Route, and
NN algorithm.